Промежуточный коммит, не рабочая версия

+++ app/src/main/java/ru/gb/makulin/mymoviefinder/facade/MovieDTO.kt
@@ -3,15 +3,15 @@ package ru.gb.makulin.mymoviefinder.facade
 data class MovieDTO(
 val adult:Boolean,
 val genres: List<GenreDTO>,
-val id: Long,
+val id: Int,
 val original_title: String,
 val title: String,
 val overview: String,
 val popularity: Double,
 val poster_path: String,
 val release_date: String,
-val runtime: Long,
+val runtime: Int,
 val tagline: String,
 val vote_average: Double,
-val vote_count: Long
+val vote_count: Int
 )
